INTG-1516 implemented multi agent v2
