--- Runtime/Scripts/Data/Packets/ConversationPacket.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 221b8d9ea161417ea2dcbcabd992a634
-timeCreated: 1713816227
